pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "django-bootstrap5>=25.1",
    "django>=5.2",
    "django-crispy-forms>=2.3",
    "pillow>=11.1.0",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.0",
    "django-allauth>=65.7.0",
    "django-cookie-consent>=0.6.0",
    "requests>=2.32.3",
    "pyjwt>=2.10.1",
    "cryptography>=44.0.2",
]
